Add aweXpect.Reflection.Internal.Tests project
