+++ Tests/aweXpect.Reflection.Internal.Tests/Extensions/ConstructorInfoExtensionsTests.cs
@@ -1,14 +1,18 @@
-﻿using System.Reflection;
+﻿using System.Linq;
+using System.Reflection;
 using aweXpect.Reflection.Extensions;
 
-namespace aweXpect.Reflection.Tests.Extensions;
+namespace aweXpect.Reflection.Internal.Tests.Extensions;
 
-public sealed class FieldInfoExtensionsTests
+// ReSharper disable UnusedParameter.Local
+// ReSharper disable UnusedMember.Local
+public sealed class ConstructorInfoExtensionsTests
 {
 	[Fact]
 	public async Task HasAttribute_WithAttribute_ShouldReturnTrue()
 	{
-		FieldInfo type = typeof(TestClass).GetField(nameof(TestClass.Field1WithAttribute))!;
+		ConstructorInfo type = typeof(TestClass).GetDeclaredConstructors()
+			.Single(c => c.GetParameters().Length == 2);
 
 		bool result = type.HasAttribute<DummyAttribute>();
 
@@ -18,7 +22,8 @@ public sealed class FieldInfoExtensionsTests
 	[Fact]
 	public async Task HasAttribute_WithoutAttribute_ShouldReturnFalse()
 	{
-		FieldInfo type = typeof(TestClass).GetField(nameof(TestClass.Field2WithoutAttribute))!;
+		ConstructorInfo type = typeof(TestClass).GetDeclaredConstructors()
+			.Single(c => c.GetParameters().Length == 3);
 
 		bool result = type.HasAttribute<DummyAttribute>();
 
@@ -28,7 +33,8 @@ public sealed class FieldInfoExtensionsTests
 	[Fact]
 	public async Task HasAttribute_WithPredicate_ShouldReturnPredicateResult()
 	{
-		FieldInfo type = typeof(TestClass).GetField(nameof(TestClass.Field1WithAttribute))!;
+		ConstructorInfo type = typeof(TestClass).GetDeclaredConstructors()
+			.Single(c => c.GetParameters().Length == 2);
 
 		bool result1 = type.HasAttribute<DummyAttribute>(d => d.Value == 1);
 		bool result2 = type.HasAttribute<DummyAttribute>(d => d.Value == 2);
@@ -37,7 +43,7 @@ public sealed class FieldInfoExtensionsTests
 		await That(result2).IsFalse();
 	}
 
-	[AttributeUsage(AttributeTargets.Field)]
+	[AttributeUsage(AttributeTargets.Constructor)]
 	private class DummyAttribute : Attribute
 	{
 		public DummyAttribute(int value)
@@ -48,12 +54,15 @@ public sealed class FieldInfoExtensionsTests
 		public int Value { get; }
 	}
 
-#pragma warning disable CS0649
 	private class TestClass
 	{
-		[Dummy(1)] public int Field1WithAttribute = 1;
+		[Dummy(1)]
+		public TestClass(int value1, int value2)
+		{
+		}
 
-		public int Field2WithoutAttribute = 1;
+		public TestClass(int value1, int value2, int value3)
+		{
+		}
 	}
-#pragma warning restore CS0649
 }
