+++ Tests/aweXpect.Reflection.Internal.Tests/Usings.cs
@@ -1,0 +1,5 @@
+﻿global using System;
+global using System.Threading.Tasks;
+global using Xunit;
+global using aweXpect;
+global using static aweXpect.Expect;
