+++ aweXpect.Reflection.sln
@@ -51,6 +51,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aweXpect.Reflection.Tests",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aweXpect.Reflection.Api.Tests", "Tests\aweXpect.Reflection.Api.Tests\aweXpect.Reflection.Api.Tests.csproj", "{EA828A6C-5C01-4950-9F67-F53847CE3DDF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aweXpect.Reflection.Internal.Tests", "Tests\aweXpect.Reflection.Internal.Tests\aweXpect.Reflection.Internal.Tests.csproj", "{13CDAE88-D6C0-45DA-B948-96088AC775B1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -65,6 +67,7 @@ Global
 		{E122DD93-09CD-4316-A0D0-F57019A3B9B5} = {B776D4D9-F241-49BA-B005-8484BC686008}
 		{494F4BA5-02EE-4656-9D5E-BF9BB25C49E8} = {9CC57AD0-4984-4618-96EA-01FFFCCD84FA}
 		{EA828A6C-5C01-4950-9F67-F53847CE3DDF} = {9CC57AD0-4984-4618-96EA-01FFFCCD84FA}
+		{13CDAE88-D6C0-45DA-B948-96088AC775B1} = {9CC57AD0-4984-4618-96EA-01FFFCCD84FA}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{E122DD93-09CD-4316-A0D0-F57019A3B9B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -83,5 +86,9 @@ Global
 		{EA828A6C-5C01-4950-9F67-F53847CE3DDF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EA828A6C-5C01-4950-9F67-F53847CE3DDF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EA828A6C-5C01-4950-9F67-F53847CE3DDF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{13CDAE88-D6C0-45DA-B948-96088AC775B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{13CDAE88-D6C0-45DA-B948-96088AC775B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{13CDAE88-D6C0-45DA-B948-96088AC775B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{13CDAE88-D6C0-45DA-B948-96088AC775B1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
